Revert sky changes
