+++ site/routes/definitions.js
@@ -132,17 +132,9 @@ export const routeDefinitions: Array<RouteDefinition> = [
     }),
   },
   {
-    title: 'Sky CMS case study',
-    key: 'skyCMSCaseStudy',
-    route: 'our-work/case-study/sky-cms',
-    stateToProps: ({ contactUsURL }) => ({
-      contactUsURL,
-    }),
-  },
-  {
     title: 'Sky case study',
     key: 'skyCaseStudy',
-    route: 'our-work/case-study/sky',
+    route: 'our-work/case-study/sky-cms',
     stateToProps: ({ contactUsURL }) => ({
       contactUsURL,
     }),
