+++ site/routes/index.js
@@ -22,7 +22,6 @@ import Technology from '../pages/technology';
 import RetailerCaseStudy from '../pages/our-work/case-study/retailer';
 import CamdenMarketCaseStudy from '../pages/our-work/case-study/camden-market';
 import FinancialTimesCaseStudy from '../pages/our-work/case-study/ft';
-import SkyCMSCaseStudy from '../pages/our-work/case-study/sky-cms';
 import SkyCaseStudy from '../pages/our-work/case-study/sky';
 import BMWCaseStudy from '../pages/our-work/case-study/bmw';
 import BBCCaseStudy from '../pages/our-work/case-study/bbc';
@@ -45,7 +44,6 @@ const componentMap = {
   retailerCaseStudy: RetailerCaseStudy,
   camdenMarketCaseStudy: CamdenMarketCaseStudy,
   financialTimesCaseStudy: FinancialTimesCaseStudy,
-  skyCMSCaseStudy: SkyCMSCaseStudy,
   skyCaseStudy: SkyCaseStudy,
   bmwCaseStudy: BMWCaseStudy,
   bbcCaseStudy: BBCCaseStudy,
